Remove old twi0_interface files. Create light_sens_interface, prox_sens_interface, fc_interface  files. UNTESTED
