+++ RobotNoASF/Debug/Makefile
@@ -45,8 +45,10 @@ C_SRCS +=  \
 ../motor_driver.c \
 ../opt_interface.c \
 ../imu_interface.c \
+../prox_sens_interface.c \
 ../robotdefines.c \
 ../twi0_interface.c \
+../twimux_interface.c \
 ../Device_Startup/startup_sam4n.c \
 ../Device_Startup/system_sam4n.c \
 ../main.c
@@ -65,8 +67,10 @@ IMU-DMP/inv_mpu_dmp_motion_driver_CUSTOM.o \
 motor_driver.o \
 opt_interface.o \
 imu_interface.o \
+prox_sens_interface.o \
 robotdefines.o \
 twi0_interface.o \
+twimux_interface.o \
 Device_Startup/startup_sam4n.o \
 Device_Startup/system_sam4n.o \
 main.o
@@ -78,8 +82,10 @@ IMU-DMP/inv_mpu_dmp_motion_driver_CUSTOM.o \
 motor_driver.o \
 opt_interface.o \
 imu_interface.o \
+prox_sens_interface.o \
 robotdefines.o \
 twi0_interface.o \
+twimux_interface.o \
 Device_Startup/startup_sam4n.o \
 Device_Startup/system_sam4n.o \
 main.o
@@ -91,8 +97,10 @@ IMU-DMP/inv_mpu_dmp_motion_driver_CUSTOM.d \
 motor_driver.d \
 opt_interface.d \
 imu_interface.d \
+prox_sens_interface.d \
 robotdefines.d \
 twi0_interface.d \
+twimux_interface.d \
 Device_Startup/startup_sam4n.d \
 Device_Startup/system_sam4n.d \
 main.d
@@ -104,8 +112,10 @@ IMU-DMP/inv_mpu_dmp_motion_driver_CUSTOM.d \
 motor_driver.d \
 opt_interface.d \
 imu_interface.d \
+prox_sens_interface.d \
 robotdefines.d \
 twi0_interface.d \
+twimux_interface.d \
 Device_Startup/startup_sam4n.d \
 Device_Startup/system_sam4n.d \
 main.d
@@ -151,23 +161,27 @@ LINKER_SCRIPT_DEP+=  \
 
 
 
+
+
+
+
 ./%.o: .././%.c
 	@echo Building file: $<
-	@echo Invoking: ARM/GNU C Compiler : 5.3.1
+	@echo Invoking: ARM/GNU C Compiler : 6.2.1
 	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\arm\arm-gnu-toolchain\bin\arm-none-eabi-gcc.exe$(QUOTE)  -x c -mthumb -D__SAM4N8C__ -DDEBUG -DMPL_LOG_NDEBUG=1 -DEMPL -DUSE_DMP -DMPU9250  -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\arm\cmsis\4.2.0\CMSIS\Include" -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\Atmel\SAM4N_DFP\1.0.49\include"  -O0 -ffunction-sections -mlong-calls -g3 -Wall -mcpu=cortex-m4 -c -std=gnu99 -MD -MP -MF "$(@:%.o=%.d)" -MT"$(@:%.o=%.d)" -MT"$(@:%.o=%.o)"   -o "$@" "$<" 
 	@echo Finished building: $<
 	
 
 IMU-DMP/%.o: ../IMU-DMP/%.c
 	@echo Building file: $<
-	@echo Invoking: ARM/GNU C Compiler : 5.3.1
+	@echo Invoking: ARM/GNU C Compiler : 6.2.1
 	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\arm\arm-gnu-toolchain\bin\arm-none-eabi-gcc.exe$(QUOTE)  -x c -mthumb -D__SAM4N8C__ -DDEBUG -DMPL_LOG_NDEBUG=1 -DEMPL -DUSE_DMP -DMPU9250  -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\arm\cmsis\4.2.0\CMSIS\Include" -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\Atmel\SAM4N_DFP\1.0.49\include"  -O0 -ffunction-sections -mlong-calls -g3 -Wall -mcpu=cortex-m4 -c -std=gnu99 -MD -MP -MF "$(@:%.o=%.d)" -MT"$(@:%.o=%.d)" -MT"$(@:%.o=%.o)"   -o "$@" "$<" 
 	@echo Finished building: $<
 	
 
 Device_Startup/%.o: ../Device_Startup/%.c
 	@echo Building file: $<
-	@echo Invoking: ARM/GNU C Compiler : 5.3.1
+	@echo Invoking: ARM/GNU C Compiler : 6.2.1
 	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\arm\arm-gnu-toolchain\bin\arm-none-eabi-gcc.exe$(QUOTE)  -x c -mthumb -D__SAM4N8C__ -DDEBUG -DMPL_LOG_NDEBUG=1 -DEMPL -DUSE_DMP -DMPU9250  -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\arm\cmsis\4.2.0\CMSIS\Include" -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\Atmel\SAM4N_DFP\1.0.49\include"  -O0 -ffunction-sections -mlong-calls -g3 -Wall -mcpu=cortex-m4 -c -std=gnu99 -MD -MP -MF "$(@:%.o=%.d)" -MT"$(@:%.o=%.d)" -MT"$(@:%.o=%.o)"   -o "$@" "$<" 
 	@echo Finished building: $<
 	
@@ -196,7 +210,7 @@ all: $(OUTPUT_FILE_PATH) $(ADDITIONAL_DEPENDENCIES)
 
 $(OUTPUT_FILE_PATH): $(OBJS) $(USER_OBJS) $(OUTPUT_FILE_DEP) $(LIB_DEP) $(LINKER_SCRIPT_DEP)
 	@echo Building target: $@
-	@echo Invoking: ARM/GNU Linker : 5.3.1
+	@echo Invoking: ARM/GNU Linker : 6.2.1
 	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\arm\arm-gnu-toolchain\bin\arm-none-eabi-gcc.exe$(QUOTE) -o$(OUTPUT_FILE_PATH_AS_ARGS) $(OBJS_AS_ARGS) $(USER_OBJS) $(LIBS) -mthumb -Wl,-Map="RobotNoASF.map" -Wl,--start-group -lm  -Wl,--end-group -L"..\\Device_Startup"  -Wl,--gc-sections -mcpu=cortex-m4 -Tsam4n8c_flash.ld  
 	@echo Finished building target: $@
 	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\arm\arm-gnu-toolchain\bin\arm-none-eabi-objcopy.exe" -O binary "RobotNoASF.elf" "RobotNoASF.bin"
