--- RobotNoASF/twi0_interface.h
@@ -1,28 +1,0 @@
-/*
- * twi0_interface.h
- *
- * Created: 6/07/2017 12:06:55 p.m.
- *  Author: Adam Parlane
- */ 
-
-//This file will contain all the relevant setup and functions relating to 
-//TWI mux, proximity, line, light sensors
-
-
-#ifndef TWI0_INTERFACE_H_
-#define TWI0_INTERFACE_H_
-
-#include "sam.h"
-#include "robotsetup.h"
-
-
-
-
-
-
-
-
-
-
-
-#endif /* TWI0_INTERFACE_H_ */
